Cleaned and documented more methods and players
