+++ src/players/MiniMaxPlayer.py
@@ -1,19 +1,20 @@
-import math
-from typing import Optional, Union
-
-import numpy as np
 from poke_env import PlayerConfiguration, ServerConfiguration
+from poke_env.environment import Status, Gen8Move
 from poke_env.player import Player
 from poke_env.teambuilder import Teambuilder
-
 from src.utilities.BattleStatus import BattleStatus
 from src.utilities.Heuristic import Heuristic
 from src.utilities.NodePokemon import NodePokemon
 from src.engine.battle_utilities import *
 from src.engine.stats import compute_stat
+from src.strategy.gimmick import should_dynamax
+from src.strategy.matchup import matchup_on_types
+from src.strategy.switch import should_switch, compute_best_switch
 from src.utilities.SimpleHeuristic import SimpleHeuristic
 from src.utilities.utilities import matchups_to_string
 from src.engine.damage import compute_damage
+from typing import Optional, Union
+import math
 
 
 class MiniMaxPlayer(Player):
@@ -30,7 +31,7 @@ class MiniMaxPlayer(Player):
                  max_concurrent_battles: int = 1,
                  save_replays: Union[bool, str] = False,
                  server_configuration: Optional[ServerConfiguration] = None,
-                 start_timer_on_battle_start: bool = False,
+                 # start_timer_on_battle_start: bool = False,
                  start_listening: bool = True,
                  ping_interval: Optional[float] = 20.0,
                  ping_timeout: Optional[float] = 20.0,
@@ -75,10 +76,12 @@ class MiniMaxPlayer(Player):
         best_switch, bot_matchup, outspeed_p, team_matchups = self.best_switch_on_matchup(battle, bot_pokemon, bot_team,
                                                                                           opp_pokemon, terrains,
                                                                                           weather)
-        if self.__should_switch(bot_pokemon, bot_matchup, outspeed_p) and battle.available_switches:
+        if should_switch(bot_pokemon, bot_matchup, outspeed_p, self.max_team_matchup, self.toxic_turn)\
+                and battle.available_switches:
             self.previous_pokemon = bot_pokemon
             if self.verbose:
                 print("Switching to {0}\n{1}".format(best_switch.species, "-" * 110))
+
             return self.create_order(best_switch)
 
         if battle.available_moves:
@@ -97,24 +100,29 @@ class MiniMaxPlayer(Player):
             can_defeat, best_move = False, Gen8Move('splash')
             if root_battle_status.move_first and len(battle.available_moves) > 0:
                 can_defeat, best_move = self.hit_if_act_poke_can_outspeed(battle, terrains, opp_max_hp, opp_conditions)
+
             if len(battle.available_moves) == 0 or can_defeat is not True:
                 best_move = self.get_best_move(battle, root_battle_status)
+
             dynamax: bool = False
             my_team = [poke for poke in list(battle.team.values()) if poke.status != Status.FNT and not poke.active]
             if battle.can_dynamax and not isinstance(best_move, Pokemon):
-                dynamax = self.should_dynamax(battle.active_pokemon, my_team, bot_matchup)
-            if self.verbose: self.print_chosen_move(battle, best_move, opp_conditions, terrains, weather)
+                dynamax = should_dynamax(battle.active_pokemon, my_team, bot_matchup,
+                                         self.max_team_matchup, self.best_stats_pokemon)
+
+            if self.verbose:
+                self.print_chosen_move(battle, best_move, opp_conditions, terrains, weather)
 
             return self.create_order(best_move, dynamax=dynamax)
 
         elif battle.available_switches:
             # Update the matchup for each remaining pokèmon in the team
             for pokemon in bot_team:
-                team_matchups.update({pokemon: self.__matchup_on_types(pokemon, opp_pokemon)})
+                team_matchups.update({pokemon: matchup_on_types(pokemon, opp_pokemon)})
 
             # Choose the new active pokèmon
             self.max_team_matchup = max(team_matchups.values()) if len(team_matchups) > 0 else -8
-            best_switch = self.__compute_best_switch(team_matchups, opp_pokemon, weather, terrains)
+            best_switch = compute_best_switch(team_matchups, opp_pokemon, weather, terrains, self.max_team_matchup)
             self.previous_pokemon = bot_pokemon
             if self.verbose:
                 print("Switching to {0}\n{1}".format(best_switch.species, "-" * 110))
@@ -125,13 +133,15 @@ class MiniMaxPlayer(Player):
 
     def best_switch_on_matchup(self, battle, bot_pokemon, bot_team, opp_pokemon, terrains, weather):
         # Compute matchup scores for every remaining pokémon in the team
-        bot_matchup = self.__matchup_on_types(bot_pokemon, opp_pokemon)
+        bot_matchup = matchup_on_types(bot_pokemon, opp_pokemon)
         team_matchups = dict()
         for pokemon in bot_team:
-            team_matchups.update({pokemon: self.__matchup_on_types(pokemon, opp_pokemon)})
+            team_matchups.update({pokemon: matchup_on_types(pokemon, opp_pokemon)})
+
         # Set the best pokémon in terms of stats
         if battle.turn == 1:
             self.best_stats_pokemon = max([sum(pokemon.base_stats.values()) for pokemon in battle.team.values()])
+
         # If we switched pokémon, then update the bot's infos
         if not self.previous_pokemon:
             self.previous_pokemon = bot_pokemon
@@ -142,9 +152,11 @@ class MiniMaxPlayer(Player):
             bot_pokemon._first_turn = False
             if bot_pokemon.status is Status.TOX:
                 self.toxic_turn += 1
+
         # Compute the best pokémon the bot can switch to
         self.max_team_matchup = max(team_matchups.values()) if len(team_matchups) > 0 else -8
-        best_switch = self.__compute_best_switch(team_matchups, opp_pokemon, weather, terrains)
+        best_switch = compute_best_switch(team_matchups, opp_pokemon, weather, terrains, self.max_team_matchup)
+
         # Compute the probability of outpseeding the opponent pokémon
         outspeed_p, opp_spe_lb, opp_spe_ub = outspeed_prob(bot_pokemon, opp_pokemon, weather, terrains).values()
         if self.verbose and bot_pokemon.current_hp != 0:
@@ -154,9 +166,10 @@ class MiniMaxPlayer(Player):
 
         return best_switch, bot_matchup, outspeed_p, team_matchups
 
-    def choose_move_old(self, battle):
+    '''def choose_move_old(self, battle):
         if battle.turn == 1:
             self.best_stats_pokemon = max([sum(pokemon.base_stats.values()) for pokemon in battle.team.values()])
+        
         weather, terrains, bot_conditions, opp_conditions = retrieve_battle_status(battle).values()
         opp_max_hp = compute_stat(battle.opponent_active_pokemon, "hp", weather, terrains)
         opp_team = [poke for poke in battle.opponent_team.values() if poke.status != Status.FNT and not poke.active]
@@ -180,11 +193,14 @@ class MiniMaxPlayer(Player):
         my_team = [poke for poke in list(battle.team.values()) if poke.status != Status.FNT and not poke.active]
         if battle.can_dynamax and not isinstance(best_move, Pokemon):
             dynamax = self.should_dynamax(battle.active_pokemon, my_team)
-        if self.verbose: self.print_chosen_move(battle, best_move, opp_conditions, terrains, weather)
+        
+        if self.verbose:
+            self.print_chosen_move(battle, best_move, opp_conditions, terrains, weather)
 
-        return self.create_order(best_move, dynamax=dynamax)
+        return self.create_order(best_move, dynamax=dynamax)'''
 
-    def hit_if_act_poke_can_outspeed(self, battle: AbstractBattle, terrains: list[Field], opp_max_hp: int,
+    @staticmethod
+    def hit_if_act_poke_can_outspeed(battle: AbstractBattle, terrains: list[Field], opp_max_hp: int,
                                      opp_conditions: list) -> tuple[bool, Move]:
         opp_hp = math.ceil(opp_max_hp * battle.opponent_active_pokemon.current_hp_fraction)
         for move in battle.available_moves:
@@ -198,30 +214,6 @@ class MiniMaxPlayer(Player):
             if opp_is_fainted:
                 return True, move
         return False, Gen8Move('splash')
-
-    def should_dynamax(self, bot_pokemon: Pokemon, bot_team: list[Pokemon], matchup: float) -> bool:
-        # If the pokèmon is the last one alive, use the dynamax
-        if len(bot_team) == 0:
-            return True
-
-        # If the current pokèmon is the best one in terms of base stats and the matchup is favorable, then dynamax
-        if sum(bot_pokemon.base_stats.values()) == self.best_stats_pokemon \
-                and matchup >= 1 and bot_pokemon.current_hp_fraction >= 0.8:
-            return True
-
-        if bot_pokemon.current_hp_fraction == 1:
-            # If the current pokèmon is the last one at full hp, use dynamax on it
-            if len([pokemon for pokemon in bot_team if pokemon.current_hp_fraction == 1]) == 0:
-                return True
-            else:
-                # If the current matchup is the best one, and it's favourable, then dynamax
-                if matchup >= self.max_team_matchup and matchup > 2:
-                    return True
-
-        if matchup >= self.max_team_matchup and matchup > 2 and bot_pokemon.current_hp_fraction == 1:
-            return True
-
-        return False
 
     @staticmethod
     def print_chosen_move(battle, best_move, opp_conditions, terrains, weather):
@@ -252,101 +244,11 @@ class MiniMaxPlayer(Player):
                 curr_node = curr_node.ancestor
         return best_move
 
-    @staticmethod
-    def __type_advantage(attacker: Pokemon, defender: Pokemon) -> float:
-        type_gain = max([defender.damage_multiplier(attacker_type)
-                         for attacker_type in attacker.types if attacker_type is not None])
-        return type_gain
-
-    @staticmethod
-    def __move_type_advantage(bot_pokemon: Pokemon, opponent_pokemon: Pokemon, opponent_type_adv: float) -> float:
-        # Consider the bot move-type match-up
-        bot_type_gain = [opponent_pokemon.damage_multiplier(move_bot)
-                         for move_bot in bot_pokemon.moves.values() if
-                         move_bot.category is not MoveCategory.STATUS]
-        bot_type_gain = max(bot_type_gain) if len(bot_type_gain) > 0 else 1
-        # Consider the opponent move-type match-up
-        opponent_type_gain = 0
-        if len(opponent_pokemon.moves) > 0:
-            for move_opp in opponent_pokemon.moves.values():
-                if move_opp.category is not MoveCategory.STATUS:
-                    opponent_type_gain_iter = bot_pokemon.damage_multiplier(move_opp)
-                    if opponent_type_gain_iter > opponent_type_gain:
-                        opponent_type_gain = opponent_type_gain_iter
-
-        if opponent_type_gain < opponent_type_adv and len(opponent_pokemon.moves) < 4:
-            return bot_type_gain - opponent_type_adv
-
-        return bot_type_gain - opponent_type_gain
-
-    def __should_switch(self, bot_pokemon: Pokemon, matchup: float, outspeed_p: float) -> bool:
-        # Do not switch out the pokèmon if it is dynamaxed unless there is a very bad matchup
-        if bot_pokemon.is_dynamaxed:
-            if matchup <= -4:
-                return True
-
-            return False
-
-        # We can switch out if there are matchups better than the current one
-        if self.max_team_matchup > matchup:
-            # The "toxic" status is one of the worst status, we need to attenuate its effects by switching
-            if bot_pokemon.status is Status.TOX and matchup - self.toxic_turn <= -2:
-                return True
-
-            # If one of the defense stat was decreased too much, switch out to not take heavy hits
-            if bot_pokemon.boosts["def"] <= -2 or bot_pokemon.boosts["spd"] <= -2:
-                return True
-
-            # If the attack stat was decreased too much and the pokèmon is a physical attacker, switch out
-            if bot_pokemon.base_stats["atk"] > bot_pokemon.boosts["spa"]:
-                if matchup <= -1.5 or bot_pokemon.boosts["atk"] <= -2:
-                    return True
-
-            # If the special attack stat was decreased too much and the pokèmon is a special attacker, switch out
-            if bot_pokemon.base_stats["spa"] > bot_pokemon.boosts["atk"]:
-                if matchup <= -1.5 or bot_pokemon.boosts["spa"] <= -2:
-                    return True
-
-            # Switch out if the matchup is on the opponent's favor
-            if matchup <= -1.5:
-                return True
-            elif matchup <= -1 and outspeed_p <= 0.5:
-                return True
-
-        return False
-
-    def __compute_best_switch(self,
-                              team_matchups: dict[Pokemon, float],
-                              opp_pokemon: Pokemon,
-                              weather: Weather,
-                              terrains: list[Field]) -> Union[Pokemon | None]:
-        if team_matchups:
-            # Retrieve all the pokémon in the team with the best matchup
-            best_switches = {pokemon: pokemon.stats for pokemon, matchup in team_matchups.items()
-                             if matchup == self.max_team_matchup}
-
-            # Keep those that can outspeed the opponent's pokémon
-            outspeed_opp = [pokemon for pokemon, _ in best_switches.items()
-                            if outspeed_prob(pokemon, opp_pokemon, weather, terrains)["outspeed_p"] > 0.6]
-            if outspeed_opp:
-                switch_choice = np.random.randint(0, len(outspeed_opp))
-                return outspeed_opp[switch_choice]
-            else:
-                best_switches = list(best_switches.keys())
-                switch_choice = np.random.randint(0, len(best_switches))
-                return best_switches[switch_choice]
-        else:
-            return None
-
-    def __matchup_on_types(self, bot_pokemon: Pokemon, opponent_pokemon: Pokemon) -> float:
-        bot_type_adv = self.__type_advantage(bot_pokemon, opponent_pokemon)
-        opponent_type_adv = self.__type_advantage(opponent_pokemon, bot_pokemon)
-        poke_adv = bot_type_adv - opponent_type_adv
-        move_adv = self.__move_type_advantage(bot_pokemon, opponent_pokemon, opponent_type_adv)
-        return poke_adv + move_adv
-
-    def alphabeta(self, node: BattleStatus, depth: int, alpha: float, beta: float, is_my_turn: bool) -> tuple[
-        float, BattleStatus]:
+    def alphabeta(self, node: BattleStatus,
+                  depth: int,
+                  alpha: float,
+                  beta: float,
+                  is_my_turn: bool) -> tuple[float, BattleStatus]:
         """
         (* Initial call *) alphabeta(origin, 0, −inf, +inf, TRUE)
         """
