+++ src/players/RuleBasedPlayer.py
@@ -5,10 +5,12 @@ from src.engine.move_effects import compute_healing
 from src.engine.battle_utilities import *
 from src.engine.useful_data import *
 from src.engine.stats import compute_stat
-from src.utilities.utilities import matchups_to_string
 from src.engine.damage import compute_damage
+from src.strategy.gimmick import should_dynamax
+from src.strategy.matchup import matchup_on_types
+from src.strategy.switch import should_switch, compute_best_switch
+from src.utilities.utilities import matchups_to_string
 from typing import Optional, Union
-import numpy as np
 
 
 class RuleBasedPlayer(Player):
@@ -22,7 +24,7 @@ class RuleBasedPlayer(Player):
                  max_concurrent_battles: int = 1,
                  save_replays: Union[bool, str] = False,
                  server_configuration: Optional[ServerConfiguration] = None,
-                 start_timer_on_battle_start: bool = False,
+                 # start_timer_on_battle_start: bool = False,
                  start_listening: bool = True,
                  ping_interval: Optional[float] = 20.0,
                  ping_timeout: Optional[float] = 20.0,
@@ -40,124 +42,6 @@ class RuleBasedPlayer(Player):
         self.previous_pokemon = None
         self.max_team_matchup = -8
         self.toxic_turn = 0
-
-    @staticmethod
-    def __type_advantage(attacker: Pokemon, defender: Pokemon) -> float:
-        type_gain = max([defender.damage_multiplier(attacker_type)
-                         for attacker_type in attacker.types if attacker_type is not None])
-        return type_gain
-
-    @staticmethod
-    def __move_type_advantage(bot_pokemon: Pokemon, opponent_pokemon: Pokemon, opponent_type_adv: float) -> float:
-        # Consider the bot move-type match-up
-        bot_type_gain = [opponent_pokemon.damage_multiplier(move_bot)
-                         for move_bot in bot_pokemon.moves.values() if
-                         move_bot.category is not MoveCategory.STATUS]
-        bot_type_gain = max(bot_type_gain) if len(bot_type_gain) > 0 else 1
-
-        # Consider the opponent move-type match-up
-        opponent_type_gain = 0
-        if len(opponent_pokemon.moves) > 0:
-            for move_opp in opponent_pokemon.moves.values():
-                if move_opp.category is not MoveCategory.STATUS:
-                    opponent_type_gain_iter = bot_pokemon.damage_multiplier(move_opp)
-                    if opponent_type_gain_iter > opponent_type_gain:
-                        opponent_type_gain = opponent_type_gain_iter
-
-        if opponent_type_gain < opponent_type_adv and len(opponent_pokemon.moves) < 4:
-            return bot_type_gain - opponent_type_adv
-
-        return bot_type_gain - opponent_type_gain
-
-    def __matchup_on_types(self, bot_pokemon: Pokemon, opponent_pokemon: Pokemon) -> float:
-        bot_type_adv = self.__type_advantage(bot_pokemon, opponent_pokemon)
-        opponent_type_adv = self.__type_advantage(opponent_pokemon, bot_pokemon)
-        poke_adv = bot_type_adv - opponent_type_adv
-        move_adv = self.__move_type_advantage(bot_pokemon, opponent_pokemon, opponent_type_adv)
-        return poke_adv + move_adv
-
-    def __should_switch(self, bot_pokemon: Pokemon, matchup: float, outspeed_p: float) -> bool:
-        # Do not switch out the pokèmon if it is dynamaxed unless there is a very bad matchup
-        if bot_pokemon.is_dynamaxed:
-            if matchup <= -4:
-                return True
-
-            return False
-
-        # We can switch out if there are matchups better than the current one
-        if self.max_team_matchup > matchup:
-            # The "toxic" status is one of the worst status, we need to attenuate its effects by switching
-            if bot_pokemon.status is Status.TOX and matchup - self.toxic_turn <= -2:
-                return True
-
-            # If one of the defense stat was decreased too much, switch out to not take heavy hits
-            if bot_pokemon.boosts["def"] <= -2 or bot_pokemon.boosts["spd"] <= -2:
-                return True
-
-            # If the attack stat was decreased too much and the pokèmon is a physical attacker, switch out
-            if bot_pokemon.base_stats["atk"] > bot_pokemon.boosts["spa"]:
-                if matchup <= -1.5 or bot_pokemon.boosts["atk"] <= -2:
-                    return True
-
-            # If the special attack stat was decreased too much and the pokèmon is a special attacker, switch out
-            if bot_pokemon.base_stats["spa"] > bot_pokemon.boosts["atk"]:
-                if matchup <= -1.5 or bot_pokemon.boosts["spa"] <= -2:
-                    return True
-
-            # Switch out if the matchup is on the opponent's favor
-            if matchup <= -1.5:
-                return True
-            elif matchup <= -1 and outspeed_p <= 0.5:
-                return True
-
-        return False
-
-    def __should_dynamax(self, bot_pokemon: Pokemon, bot_team: list[Pokemon], matchup: float) -> bool:
-        # If the pokèmon is the last one alive, use the dynamax
-        if len(bot_team) == 0:
-            return True
-
-        # If the current pokèmon is the best one in terms of base stats and the matchup is favorable, then dynamax
-        if sum(bot_pokemon.base_stats.values()) == self.best_stats_pokemon \
-                and matchup >= 1 and bot_pokemon.current_hp_fraction >= 0.8:
-            return True
-
-        if bot_pokemon.current_hp_fraction == 1:
-            # If the current pokèmon is the last one at full hp, use dynamax on it
-            if len([pokemon for pokemon in bot_team if pokemon.current_hp_fraction == 1]) == 0:
-                return True
-            else:
-                # If the current matchup is the best one, and it's favourable, then dynamax
-                if matchup >= self.max_team_matchup and matchup > 2:
-                    return True
-
-        if matchup >= self.max_team_matchup and matchup > 2 and bot_pokemon.current_hp_fraction == 1:
-            return True
-
-        return False
-
-    def __compute_best_switch(self,
-                              team_matchups: dict[Pokemon, float],
-                              opp_pokemon: Pokemon,
-                              weather: Weather,
-                              terrains: list[Field]) -> Union[Pokemon | None]:
-        if team_matchups:
-            # Retrieve all the pokémon in the team with the best matchup
-            best_switches = {pokemon: pokemon.stats for pokemon, matchup in team_matchups.items()
-                             if matchup == self.max_team_matchup}
-
-            # Keep those that can outspeed the opponent's pokémon
-            outspeed_opp = [pokemon for pokemon, _ in best_switches.items()
-                            if outspeed_prob(pokemon, opp_pokemon, weather, terrains)["outspeed_p"] > 0.6]
-            if outspeed_opp:
-                switch_choice = np.random.randint(0, len(outspeed_opp))
-                return outspeed_opp[switch_choice]
-            else:
-                best_switches = list(best_switches.keys())
-                switch_choice = np.random.randint(0, len(best_switches))
-                return best_switches[switch_choice]
-        else:
-            return None
 
     @staticmethod
     def __compute_opponent_damage(bot_pokemon: Pokemon,
@@ -193,10 +77,10 @@ class RuleBasedPlayer(Player):
                     if pokemon is not bot_pokemon and not pokemon.fainted]
 
         # Compute matchup scores for every remaining pokémon in the team
-        bot_matchup = self.__matchup_on_types(bot_pokemon, opp_pokemon)
+        bot_matchup = matchup_on_types(bot_pokemon, opp_pokemon)
         team_matchups = dict()
         for pokemon in bot_team:
-            team_matchups.update({pokemon: self.__matchup_on_types(pokemon, opp_pokemon)})
+            team_matchups.update({pokemon: matchup_on_types(pokemon, opp_pokemon)})
 
         # Compute the best base stats, this is useful for dynamax
         if battle.turn == 1:
@@ -222,7 +106,7 @@ class RuleBasedPlayer(Player):
 
         # Compute the best pokémon the bot can switch to
         self.max_team_matchup = max(team_matchups.values()) if len(team_matchups) > 0 else -8
-        best_switch = self.__compute_best_switch(team_matchups, opp_pokemon, weather, terrains)
+        best_switch = compute_best_switch(team_matchups, opp_pokemon, weather, terrains, self.max_team_matchup)
 
         # Compute the hp of both pokémon
         bot_hp = bot_pokemon.current_hp
@@ -376,7 +260,8 @@ class RuleBasedPlayer(Player):
 
             # Deal with stall pokémon
             if "toxic" in available_moves_ids and len(bot_healing_moves) > 0:
-                if battle.available_switches and self.__should_switch(bot_pokemon, bot_matchup, outspeed_p):
+                if battle.available_switches and should_switch(bot_pokemon, bot_matchup, outspeed_p,
+                                                               self.max_team_matchup, self.toxic_turn):
                     self.previous_pokemon = bot_pokemon
                     if self.verbose:
                         print("\nSwitching to {0}\n{1}".format(best_switch.species, "-" * 110))
@@ -384,7 +269,8 @@ class RuleBasedPlayer(Player):
                     return self.create_order(best_switch)
 
             # Switch out if there are available switches and the bot gains in terms of matchup by doing so
-            if battle.available_switches and self.__should_switch(bot_pokemon, bot_matchup, outspeed_p):
+            if battle.available_switches and should_switch(bot_pokemon, bot_matchup, outspeed_p,
+                                                           self.max_team_matchup, self.toxic_turn):
                 self.previous_pokemon = bot_pokemon
                 if self.verbose:
                     print("\nSwitching to {0}\n{1}".format(best_switch.species, "-" * 110))
@@ -495,7 +381,8 @@ class RuleBasedPlayer(Player):
             # Use the dynamax if convenient
             dynamax = False
             if battle.can_dynamax:
-                dynamax = self.__should_dynamax(bot_pokemon, bot_team, bot_matchup)
+                dynamax = should_dynamax(bot_pokemon, bot_team, bot_matchup,
+                                         self.max_team_matchup, self.best_stats_pokemon)
 
             if self.verbose:
                 if self.verbose:
@@ -505,11 +392,11 @@ class RuleBasedPlayer(Player):
         elif battle.available_switches:
             # Update the matchup for each remaining pokèmon in the team
             for pokemon in bot_team:
-                team_matchups.update({pokemon: self.__matchup_on_types(pokemon, opp_pokemon)})
+                team_matchups.update({pokemon: matchup_on_types(pokemon, opp_pokemon)})
 
             # Choose the new active pokèmon
             self.max_team_matchup = max(team_matchups.values()) if len(team_matchups) > 0 else -8
-            best_switch = self.__compute_best_switch(team_matchups, opp_pokemon, weather, terrains)
+            best_switch = compute_best_switch(team_matchups, opp_pokemon, weather, terrains, self.max_team_matchup)
             self.previous_pokemon = bot_pokemon
             if self.verbose:
                 print("Switching to {0}\n{1}".format(best_switch.species, "-" * 110))
